Add logging

+++ server/routers/notes.py
@@ -6,9 +6,10 @@ from crud.notes import NoteService
 from crud.users import UserService
 
 from schemas.notes import Note, NoteCreation, NoteUpdate
+from services.auth import access_granted
 
 # Маршрут для заметок
-router = APIRouter(prefix='/api/v1/notes',)
+router = APIRouter(prefix='/api/v1/notes')
 
 
 # Маршрут для получения всех заметок для авторизированных пользователей
@@ -16,11 +17,11 @@ router = APIRouter(prefix='/api/v1/notes',)
             status_code=status.HTTP_200_OK, name='get_all_notes_url',
             responses={200: {'description': 'Успешное получение объектов'}, }
             )
-async def get_notes(request: Request, tg_id: int, tag_id: int = None, service: NoteService = Depends(),
-                    user_service: UserService= Depends(), ):
-
-    result = await service.get_all_notes(tg_id=tg_id, tag_id=tag_id)
-    return result
+async def get_notes(tg_id: int, tag_id: int = None, service: NoteService = Depends(),
+                    permission: bool = Depends(access_granted)):
+    if permission:
+        result = await service.get_all_notes(tg_id=tg_id, tag_id=tag_id)
+        return result
 
 
 # Маршрут для получения заметки по id для авторизированных пользователей
@@ -28,9 +29,11 @@ async def get_notes(request: Request, tg_id: int, tag_id: int = None, service: N
             status_code=status.HTTP_200_OK, name='get_note_url',
             responses={200: {'description': 'Успешное получение объекта'}, }
             )
-async def get_notes(note_id: int, tg_id: int, service: NoteService = Depends(), ):
-    result = await service.get_note_by_id(note_id=note_id, tg_id=tg_id)
-    return result
+async def get_notes(note_id: int, tg_id: int, service: NoteService = Depends(),
+                    permission: bool = Depends(access_granted)):
+    if permission:
+        result = await service.get_note_by_id(note_id=note_id, tg_id=tg_id)
+        return result
 
 
 # Маршрут для создания заметки для авторизированных пользователей
@@ -39,10 +42,11 @@ async def get_notes(note_id: int, tg_id: int, service: NoteService = Depends(), 
              responses={201: {'description': 'Успешное создание объекта'},
                         409: {'description': 'Объект уже существует'}}
              )
-async def post_note(data: NoteCreation, service: NoteService = Depends()) -> dict[str, str]:
-    result = await service.create_note(data)
-
-    return result
+async def post_note(data: NoteCreation, service: NoteService = Depends(),
+                    permission: bool = Depends(access_granted)) -> dict[str, str]:
+    if permission:
+        result = await service.create_note(data)
+        return result
 
 
 # Маршрут для удаления заметки для авторизированных пользователей
@@ -52,9 +56,11 @@ async def post_note(data: NoteCreation, service: NoteService = Depends()) -> dic
                           404: {'description': 'Объект не найден'},
                           422: {'description': 'Ошибка валидации данных'}}
                )
-async def delete_note(note_id: int, service: NoteService = Depends()) -> dict[str, str]:
-    result = await service.delete_note(note_id)
-    return result
+async def delete_note(note_id: int, service: NoteService = Depends(),
+                      permission: bool = Depends(access_granted)) -> dict[str, str]:
+    if permission:
+        result = await service.delete_note(note_id)
+        return result
 
 
 # Маршрут для обновления заметки для авторизированных пользователей
@@ -64,6 +70,8 @@ async def delete_note(note_id: int, service: NoteService = Depends()) -> dict[st
                          404: {'description': 'Объект не найден'},
                          422: {'description': 'Ошибка валидации данных'}}
               )
-async def patch_note(note_id: int, data: NoteUpdate, service: NoteService = Depends()) -> dict[str, str]:
-    result = await service.update_note(note_id=note_id, data=data)
-    return result
+async def patch_note(note_id: int, data: NoteUpdate, service: NoteService = Depends(),
+                     permission: bool = Depends(access_granted)) -> dict[str, str]:
+    if permission:
+        result = await service.update_note(note_id=note_id, data=data)
+        return result
